backend/requirements.txt
# Web framework
fastapi==0.115.12
uvicorn[standard]==0.34.0

# Database
prisma==0.15.0

# Environment
python-dotenv==1.1.0
pydantic-settings==2.8.1
email-validator==2.2.0

# Auth
python-jose[cryptography]==3.4.0
bcrypt==4.3.0
python-multipart==0.0.20
Pillow==11.1.0

# Testing
pytest==8.3.5
pytest-asyncio==0.26.0
httpx==0.28.1
